Move networking in nodeup to dedicated subpackage
