+++ nodeup/pkg/model/BUILD.bazel
@@ -4,7 +4,6 @@ go_library(
     name = "go_default_library",
     srcs = [
         "architecture.go",
-        "cilium.go",
         "cloudconfig.go",
         "containerd.go",
         "context.go",
@@ -21,14 +20,12 @@ go_library(
         "kube_apiserver_healthcheck.go",
         "kube_controller_manager.go",
         "kube_proxy.go",
-        "kube_router.go",
         "kube_scheduler.go",
         "kubectl.go",
         "kubelet.go",
         "logrotate.go",
         "manifests.go",
         "miscutils.go",
-        "network.go",
         "node_authorizer.go",
         "ntp.go",
         "packages.go",
@@ -73,7 +70,6 @@ go_library(
         "//vendor/github.com/aws/aws-sdk-go/aws/session:go_default_library",
         "//vendor/github.com/aws/aws-sdk-go/service/ec2:go_default_library",
         "//vendor/github.com/blang/semver:go_default_library",
-        "//vendor/golang.org/x/sys/unix:go_default_library",
         "//vendor/k8s.io/api/core/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/api/resource:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
